Commit #19 - Rodrigo Viluron

+++ src/main/java/com/PimientaPasion/BuenSabor/DTOs/ClienteRankingDTO.java
@@ -11,7 +11,7 @@ import lombok.*;
 public class ClienteRankingDTO {
     private String nombre;
     private String apellido;
-    private int cantidadPedidos;
+    private Long cantidadPedidos;
     private double totalPrecioPedidos;
 
 }
